refactor: change directory and file names to kebab-case
